move lifecycle into life_cycle_steps namespace
