+++ app/components/projects/settings/life_cycle_steps/show_page_header_component.html.erb
