+++ app/components/projects/settings/life_cycle_steps/step_component.html.erb
@@ -16,7 +16,7 @@
       # behavior can be seen on primer's viewbook as well -> https://view-components-storybook.eastus.cloudapp.azure.com/view-components/lookbook/inspect/primer/alpha/toggle_switch/small
       # quick fix: don't display loading indicator which is causing the bounce
       render(Primer::Alpha::ToggleSwitch.new(
-        src: toggle_project_settings_life_cycle_path(definition_id: definition.id),
+        src: toggle_project_settings_life_cycle_step_path(id: definition.id),
         csrf_token: form_authenticity_token,
         data: { test_selector: "toggle-project-life-cycle-#{definition.id}" },
         aria: { label: toggle_aria_label },
