+++ app/views/projects/settings/life_cycle_steps/index.html.erb
@@ -27,7 +27,7 @@ See COPYRIGHT and LICENSE files for more details.
 
 ++#%>
 
-<%= render(Projects::Settings::LifeCycles::ShowPageHeaderComponent.new(project: @project)) %>
+<%= render(Projects::Settings::LifeCycleSteps::ShowPageHeaderComponent.new(project: @project)) %>
 
-<%= render(Projects::Settings::LifeCycles::IndexComponent.new(project: @project,
-                                                              life_cycle_definitions: @life_cycle_definitions)) %>
+<%= render(Projects::Settings::LifeCycleSteps::IndexComponent.new(project: @project,
+                                                                  life_cycle_definitions: @life_cycle_definitions)) %>
